Memuna_portfolio/Final Portfolio/requirements.txt
flask>=2.0.1,<3.0.0
flask-mail>=0.9.1,<1.0.0
python-dotenv>=0.19.0
flask-sqlalchemy>=2.5.1
flask-admin>=1.5.8
flask-login>=0.5.0
pillow>=9.0.0
flask-migrate>=3.1.0
werkzeug>=2.0.0
email_validator
flask-wtf>=1.0.0  # Added this line
wtforms>=3.0.0    # Added this line